@@ -9,28 +9,30 @@
       "testsFiles": "src/*.mocha.js"
     },
     "configs": [
-      "jsdocs",
-      "jsarch",
       "readme",
-      "travis",
+      "mocha",
       "eslint",
+      "travis",
       "codeclimate",
-      "mocha"
+      "babel",
+      "jsdocs",
+      "jsarch"
     ]
   },
   "scripts": {
     "architecture": "jsarch src/*.js > ARCHITECTURE.md",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "cli": "env NODE_ENV=${NODE_ENV:-cli}",
-    "cover": "istanbul cover _mocha --report html -- src/*.mocha.js -R spec -t 5000",
-    "coveralls": "istanbul cover _mocha --report lcovonly -- src/*.mocha.js -R spec -t 5000 && cat ./coverage/lcov.info | coveralls && rm -rf ./coverage",
+    "compile": "babel src --out-dir=dist",
+    "cover": "istanbul cover _mocha --report html -- --compilers js:babel-register src/*.mocha.js -R spec -t 5000",
+    "coveralls": "istanbul cover _mocha --report lcovonly -- --compilers js:babel-register src/*.mocha.js -R spec -t 5000 && cat ./coverage/lcov.info | coveralls && rm -rf ./coverage",
     "cz": "env NODE_ENV=${NODE_ENV:-cli} git cz",
-    "doc": "mkdir -p .readme; echo \"# API\" > .readme/API.md; jsdoc2md src/*.js >> .readme/API.md",
+    "doc": "echo \"# API\" > API.md; jsdoc2md src/*.js >> API.md",
     "lint": "eslint src/*.js",
     "metapak": "metapak",
-    "mocha": "mocha src/*.mocha.js",
+    "mocha": "mocha --compilers js:babel-register src/*.mocha.js",
     "prettier": "prettier --write src/*.js",
-    "preversion": "npm t && npm run lint && npm run metapak -s",
+    "preversion": "npm t && npm run lint && npm run metapak -s && npm run compile",
     "test": "npm run mocha",
     "version": "npm run changelog && git add CHANGELOG.md"
   },
@@ -45,25 +47,30 @@
   },
   "homepage": "https://github.com/nfroidure/memory-kv-store#readme",
   "dependencies": {
-    "common-services": "^1.0.0",
-    "debug": "3.1.0",
-    "knifecycle": "^2.4.2"
+    "common-services": "^3.0.1",
+    "knifecycle": "^2.5.2"
   },
   "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.2",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-register": "^6.26.0",
     "commitizen": "^2.9.6",
-    "conventional-changelog-cli": "^1.3.5",
-    "coveralls": "^2.13.3",
-    "cz-conventional-changelog": "^2.0.0",
-    "eslint": "^4.12.1",
-    "eslint-plugin-prettier": "^2.3.1",
-    "istanbul": "^0.4.5",
-    "jsarch": "^1.2.5",
-    "jsdoc-to-markdown": "^3.0.2",
-    "metapak": "^1.0.1",
-    "metapak-nfroidure": "^2.0.2",
-    "mocha": "^3.5.3",
+    "conventional-changelog-cli": "^1.3.8",
+    "coveralls": "^3.0.0",
+    "cz-conventional-changelog": "^2.1.0",
+    "eslint": "^4.19.0",
+    "eslint-plugin-prettier": "^2.6.0",
+    "istanbul": "^1.0.0-alpha.2",
+    "jsarch": "^1.3.0",
+    "jsdoc-to-markdown": "^4.0.1",
+    "metapak": "^1.0.3",
+    "metapak-nfroidure": "^6.1.0",
+    "mocha": "^5.0.0",
     "mocha-lcov-reporter": "^1.3.0",
-    "prettier": "^1.8.2",
+    "prettier": "^1.11.1",
     "sinon": "^2.1.0"
   },
   "engines": {
@@ -79,15 +86,30 @@
       "commitizen",
       "cz-conventional-changelog",
       "conventional-changelog-cli",
-      "jsdoc-to-markdown",
-      "jsarch",
-      "eslint",
-      "eslint-config-prettier",
-      "prettier",
       "mocha",
       "mocha-lcov-reporter",
       "coveralls",
-      "istanbul"
+      "istanbul",
+      "eslint",
+      "eslint-config-prettier",
+      "prettier",
+      "jsdoc-to-markdown",
+      "jsarch"
+    ]
+  },
+  "babel": {
+    "presets": [
+      [
+        "env",
+        {
+          "targets": {
+            "node": "6.9.5"
+          }
+        }
+      ]
+    ],
+    "plugins": [
+      "transform-object-rest-spread"
     ]
   }
 }
