@@ -2,6 +2,6 @@ module.exports = {
   semi: true,
   printWidth: 80,
   singleQuote: true,
-  trailingComma: 'es5',
-  proseWrap: true,
+  trailingComma: 'all',
+  proseWrap: 'always',
 };
